requirements.txt
openai
python-dotenv
tenacity
unstructured[all-docs]